+++ consumer-smoke/src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { Ng2SmartTableModule } from '@ng-smart-table/ng-smart-table';
+import { Ng2SmartTableModule } from '@yoch/ng-smart-table';
 
 import { AppComponent } from './app.component';