+++ projects/demo/src/app/pages/examples/server/basic-example-load.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 
-import { LocalDataSource } from '@ng-smart-table/ng-smart-table';
+import { LocalDataSource } from '@yoch/ng-smart-table';
 import { BasicExampleLoadService } from './basic-example-load.service';
 
 @Component({